Attack Animation
Biggest thing is attack animation with 2 out of 3 (we could go for even more combo slots) attacks implemented.
The movement is locked during the whole attack animation and is instead simulated by changing the position of the player. This might be a bit too "powerful" since it basically adds as a teleport, but the range is quite short so it might be fine.

Also changed the way the tongue works, player must hold down mouse button for it to do anything, otherwise the tongue immediately returns. No more "tap" pulls.