Pathfinding
Main thing of the commit is that pathfinding is complete (I think). The only problem I know of is that enemies don't avoid each other, so a flying enemy will try to fly through a snake for example.

There's also some ceiling sticking animation and some assets for a snake boss that I'm working on.

I tried implementing enemy attacks and it kind of works, sometimes... so that needs to be fixed (for some reason my layermasking doesn't work).