Fix animation state behaviour bug for attacking Player
Enemies could only properly attack when there was one in the scene at a time. Now they dont use the instance but rather get the Enemy component attached to the animating enemy and use that to cause an attack on the player. Seems to work

Disclaimer: I might have broken something else when I tried to fix this, if I did I'm sorry