big boy commit
This has a lot of stuff, mostly in terms of animations. It will probably take a couple of minutes to import everything, haha...

I've finished 4 boss attacks and added damage detection to them as well as colliders that follows the attack animations, so you could theoretically do damage to the boss during them. It should be in a ready state as a prefab.

I also added a very basic air attack, it's just the slash FX with no animation at the moment.