Ground pathfinding with jumps
I implemeted a new pathfinding system since I don't know how to make own nodes for A*. The new implementation mostly works very well, but there's some instances where the enemy will "dance" on the spot because it is probably updating its path unnecessarily, which will need to be fixed but it's not that big of a deal and it's in a decent condition as is.

I left the old pathfinding so that the enemies can work without setting up new nodes everywhere. To use the new pathfinding, add the "NewGroundEnemy" tag to the enemy.