Pathfinding
In its current implementation, the pathfinding now works problemfree from my testing. The AI chooses the best best path and does it fairly fast.

The way the AI chooses the path works by looking at a node's distance to the target as well as its distance to the seeker.

I also added a variable jump strength that depends on the distance of the jump but I haven't tested that thoroughly. The AI can also jump vertical gaps now.