Area1[WIP] and refactoring
Started some work on the first area. My idea is that the wasp queen is located near the top of a giant tree, so after going along the ground for a while you'll have to scale the tree to reach her.

Also introduced a nongrappable tag so the frog can't just scale all walls infinitely.