Larva, dialoguebox and wasp queen camera
Makes larva grappable.

Also added a dialoguebox for the tutorial (and for other stuff later maybe). Messages can be sped through with return or they'll expire automatically after some time. There's also a small place for an image.

The last thing added is a very zoomed out camera for the boss. I don't know if it's a good design choice, so feedback would be appreciated. I guess it could be hard to tell if it's good without having an actual boss fight though.