Wasp Queen meachanics 75% done
First attack is done, the wasp will telegraph a big attack for 2.5 seconds and then launch it at a very high speed. If the attack strikes the floor then the floor will crumble. If the floor crumbles 5 times, it disappears and the lethal spike floor underneath is exposed. After this kind of attack, the wasp is stuck for a certain amount of time.

Things I'd like to add:
* Animations of course
* Another, smaller attack that goes from side-to-side instead.
* A kind of resource that spawns near the ceiling, the wasp can gather it in-between big attacks. The more of the resource she collects, the faster and more damaging her big attack will be. The player can prevent her from gathering the resource by grappling it down from the ceiling. The resource could also be a source of health/attackboost to the player if they break/kill it midair instead of letting it fall to the ground.