Experimental "bullet time" ability
Adds the ability to slow down time with the left control key. While slowed down the player will get an attack line that turns red if it crosses an enemy. If the player presses the attack button while the line is red, a swift teleporting attack will be executed and the player will lose all their resource points (which I'm tentatively calling "ki"). If they don't attack anything, they will still lose half the points.

The resource system isn't really implemented yet, so the ability is currently spammable (but that's nice for testing). I'm thinking the resource could also be used to regenerate HP, like 25% of max resource would heal 20% HP. So it would introduce a choice of saving the resource for the big offensive ability or using it defensively to get some HP back.

I also added a hit animation and a pool for it which animates it at the point of the attack. I'm specifically using it for the bullet time attack, but it could be used for all attacks if  you guys thinks that would look good.