PC direction change
Changed the way direction works. It basically works as it did at the very start, the frog turns based on movement instead of where the mouse is, so attacks will go in the direction the frog faces and not in the direction the mouse is. The only time the direction changes based on the mouse is when the tongue is active. It might take some getting used to if you're used to the old system but I think it's more intuitive for new players.