Player attack change and bug fix
Fixes the hitbox for attacks in the left direction, the math was a bit off.

I also made the attacks a bit faster so it feels like the player is more in control of the frog, it might be too fast so feedback would be appreciated. To balance it I also lowered the damage.

An alternative fix might be to just allow the player to move halfway through the attack or something, but that might make it look like the frog is gliding around while he's stuck in the attack animation.