Checkpoints and locked dash
Implemented checkpoints that can be placed at different location on the map. They're currently invisible but they could probably be represented by a shrine or something instead.

The checkpoints can also act as scene transition triggers by enabling the bool on the object. A spawn position in the next scene as well as the scene index also have to be provided.

The dash is now locked from the start of area1 and will be unlocked when the player gets the "Waspwing Cape" by defeating the first boss. The dash is still unlocked by default if the player spawns directly to the cave area, which is preferable if we will jump around during the demo.

I also changed the airslash sound FX and some text files.