Checkpoint sprite and animation
Local checkpoints are represented by a shrine that have a flame that lights up when they become the active checkpoint. Transition checkpoints are still invisible since the player will be teleported as soon as they trigger them.

TODO: have the checkpoints be manually activated with a key input instead of automatically activating when the player gets near them.