Upwards slash
Player can do an upwards slash if they hold up while attacking.

Also adds a sprite for the air slash.

The attack movement for the normal attacks is changed since the camera got wonky with the teleporting style of the previous movement. Now the frog will "glide" a bit before coming to a halt if they're running before they attack.

A script for parallax is also included, but parallax didn't look that great with the heavy verticality of the first level and with the background that we have, so it's not heavily featured.