assets and stuff
Adds a spike hazard and a background for the wasp queen encounter.

Some movement tuning, including wall jumps that can now scale walls. The momentum of airborne players don't die now.

Also lowers toad grunt HP a bit.

Code for a respawn is also added but it probably won't be used since it introduces a lot of headaches.